ourproduct

+++ Areas/Admin/Controllers/HomeProductController.cs
@@ -37,26 +37,31 @@ namespace Backend_Final_Project.Areas.Admin.Controllers
             if (!ModelState.IsValid)
                 return View();
 
-            if (!createhomeproductVm.Photo.ContentType.Contains("image/"))
+            if (!createhomeproductVm.Photo.ContentType.Contains("image/") && !createhomeproductVm.HoverPhoto.ContentType.Contains("image/"))
                 return View();
 
-            if (createhomeproductVm.Photo.Length / 1024 > 500)
+            if (createhomeproductVm.Photo.Length / 1024 > 500 && createhomeproductVm.HoverPhoto.Length / 1024 > 500)
                 return View();
 
             string filename = Guid.NewGuid().ToString() + "_" + createhomeproductVm.Photo.FileName;
+            string Hoverfilename = Guid.NewGuid().ToString() + "_" + createhomeproductVm.HoverPhoto.FileName;
 
             string path = Path.Combine(_environment.WebRootPath, "images", filename);
+            string Hoverpath = Path.Combine(_environment.WebRootPath, "images", Hoverfilename);
 
             using FileStream stream = new FileStream(path, FileMode.Create);
+            using FileStream Hoverstream = new FileStream(Hoverpath, FileMode.Create);
 
             await createhomeproductVm.Photo.CopyToAsync(stream);
+            await createhomeproductVm.HoverPhoto.CopyToAsync(Hoverstream);
 
             HomeProduct homeproduct = new()
             {
                 Name = createhomeproductVm.Name,
                 Rating = createhomeproductVm.Rating,
                 Price = createhomeproductVm.Price,
-                Image = filename
+                Image = filename,
+                HoverImage = Hoverfilename,
             };
             await _context.HomeProducts.AddAsync(homeproduct);
             await _context.SaveChangesAsync();
@@ -75,21 +80,18 @@ namespace Backend_Final_Project.Areas.Admin.Controllers
         public async Task<IActionResult> Delete(int id)
         {
 
-            var homeProductCount = await _context.HomeProducts.CountAsync();
-            if (homeProductCount <= 2)
-                return RedirectToAction(nameof(Index));
-
-
             var homeproduct = await _context.HomeProducts.FirstOrDefaultAsync(x => x.Id == id);
             if (homeproduct == null)
                 return View();
 
             string path = Path.Combine(_environment.WebRootPath, "images", homeproduct.Image);
+            string Hoverpath = Path.Combine(_environment.WebRootPath, "images", homeproduct.HoverImage);
 
-            if (System.IO.File.Exists(path))
+            if (System.IO.File.Exists(path) && System.IO.File.Exists(Hoverpath))
+            {
                 System.IO.File.Delete(path);
-
-            System.IO.File.Delete(path);
+                System.IO.File.Delete(Hoverpath);
+            }
 
             _context.HomeProducts.Remove(homeproduct);
             await _context.SaveChangesAsync();
@@ -110,6 +112,7 @@ namespace Backend_Final_Project.Areas.Admin.Controllers
                 Price = homeproduct.Price,
                 Rating = homeproduct.Rating,
                 Image = homeproduct.Image,
+                HoverImage = homeproduct.HoverImage,
             };
 
             return View(updateHomeProductVM);
@@ -123,26 +126,36 @@ namespace Backend_Final_Project.Areas.Admin.Controllers
             var homeproduct = await _context.HomeProducts.FirstOrDefaultAsync(x => x.Id == updateHomeProductVM.Id);
             if (homeproduct == null) return NotFound();
 
-            if (updateHomeProductVM.Photo != null)
+            if (updateHomeProductVM.Photo != null && updateHomeProductVM.HoverPhoto != null)
             {
-                if (updateHomeProductVM.Photo.ContentType.Contains("image/"))
+                if (updateHomeProductVM.Photo.ContentType.Contains("image/") && updateHomeProductVM.HoverPhoto.ContentType.Contains("image/"))
                     return View();
 
-                if (updateHomeProductVM.Photo.Length / 1024 > 500)
+                if (updateHomeProductVM.Photo.Length / 1024 > 500 && updateHomeProductVM.HoverPhoto.Length / 1024 > 500)
                     return View();
 
 
                 string filename = Guid.NewGuid().ToString() + " _ " + updateHomeProductVM.Photo.FileName ;
+                string Hoverfilename = Guid.NewGuid().ToString() + " _ " + updateHomeProductVM.HoverPhoto.FileName ;
                 string path = Path.Combine(_environment.WebRootPath, "images", filename);
+                string Hoverpath = Path.Combine(_environment.WebRootPath, "images", Hoverfilename);
 
                 using FileStream stream = new FileStream(path, FileMode.Create);
+                using FileStream Hoverstream = new FileStream(Hoverpath, FileMode.Create);
 
                 await updateHomeProductVM.Photo.CopyToAsync(stream);
+                await updateHomeProductVM.HoverPhoto.CopyToAsync(Hoverstream);
 
                 string oldPath = Path.Combine(_environment.WebRootPath, "images", homeproduct.Image);
-                if (System.IO.File.Exists(oldPath))
+                string oldHoverPath = Path.Combine(_environment.WebRootPath, "images", homeproduct.HoverImage);
+                if (System.IO.File.Exists(oldPath) && System.IO.File.Exists(oldHoverPath))
+                {
                     System.IO.File.Delete(oldPath);
+                    System.IO.File.Delete(oldHoverPath);
+                }
+                 
                 homeproduct.Image = filename;
+                homeproduct.HoverImage = Hoverfilename;
             }
 
             homeproduct.Name = updateHomeProductVM.Name;
